BFS: minimize cost of generated sentences
To complete sentences, we used a simple BFS which had the effect of
minimizing the length of sentential forms (number of symbols in the
derivation), but not the length of sentences (number of terminals).

Minimizing the number of terminals is cleaner and leads to simpler
sentences. The most efficient way to achieve that would be to do a
Dijkstra's shortest-path search, but repeating the BFS as long as a
smaller cost is simpler and sufficiently fast for now.